Delete unnecessary files and edit the structure
