--- src/main/java/com/example/Mini/file/FileCloudinaryService.java
@@ -1,29 +1,0 @@
-package com.example.Mini.file;
-
-import com.cloudinary.Cloudinary;
-import com.cloudinary.utils.ObjectUtils;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.Map;
-
-@Service
-
-public class FileCloudinaryService {
-    private final Cloudinary cloudinary;
-
-    public FileCloudinaryService(Cloudinary cloudinary) {
-        this.cloudinary = cloudinary;
-    }
-
-    public String uploadFile(MultipartFile file, String folderName) throws IOException {
-        Map<?, ?> uploadResult = cloudinary.uploader().upload(
-                file.getBytes(),
-                ObjectUtils.asMap(
-                        "folder", folderName
-                )
-        );
-        return uploadResult.get("secure_url").toString();
-    }
-}
