+++ src/main/java/com/example/Mini/repository/CartItemRepository.java
@@ -8,8 +8,5 @@ import java.util.Optional;
 
 public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
     Optional<CartItem>findByUserIdAndProductId(Integer userId, Integer productId);
-
     List<CartItem> findByUserId(Integer userId);
-
-    CartItem deleteByProductIdAndUserId(Integer productId, Integer userId);
 }
