--- src/main/java/com/example/Mini/repository/UserRepository.java
@@ -1,5 +1,0 @@
-package com.example.Mini.repository;
-
-public class UserRepository {
-
-}
