--- src/main/java/com/example/Mini/request/CreateUserRequest.java
@@ -1,4 +1,0 @@
-package com.example.Mini.request;
-
-public class CreateUserRequest {
-}
