--- src/main/java/com/example/Mini/response/UserResponse.java
@@ -1,4 +1,0 @@
-package com.example.Mini.response;
-
-public class UserResponse {
-}
