--- src/main/java/com/example/Mini/service/UserService.java
@@ -1,8 +1,0 @@
-package com.example.Mini.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-}
