+++ src/main/java/com/example/Mini/statusEnum/StatusEnum.java
@@ -1,0 +1,5 @@
+package com.example.Mini.statusEnum;
+
+public enum StatusEnum {
+    SUCCESS
+}
